+++ client/app/components/EstimationArea/_styled.js
@@ -14,7 +14,15 @@ export const StyledEstimation = styled.div`
 export const EstimationAreaButtons = styled.div`
   margin-top: 16px;
   display: flex;
-  justify-content: space-between;
+  justify-content: ${({alignment}) => alignment};
+
+  > button {
+    margin-right: 4px;
+
+    &:last-child {
+      margin-right: 0;
+    }
+  }
 `;
 
 const FlashAnimation = keyframes`
@@ -64,27 +72,21 @@ export const StyledCards = styled.div`
 `;
 
 export const StyledCard = styled.button`
-  &,
-  &:focus {
-    cursor: pointer;
-    min-width: 28vw;
-    display: block;
-    border: none;
-    outline: none;
-    background: transparent;
-    text-align: center;
-    height: 100px;
-    box-sizing: border-box;
-    font-size: 28px;
-    padding: 8px;
-    font-family: ${ZuehlkeFont};
-  }
+  cursor: pointer;
+  min-width: 28vw;
+  display: block;
+  border: none;
+  outline: none;
+  background: transparent;
+  text-align: center;
+  height: 100px;
+  box-sizing: border-box;
+  font-size: 28px;
+  padding: 8px;
+  font-family: ${ZuehlkeFont};
 
   @media ${device.desktop} {
-    &,
-    &:focus {
-      min-width: 100px;
-    }
+    min-width: 100px;
   }
 `;
 
@@ -107,16 +109,10 @@ export const StyledCardInner = styled.div`
 `;
 
 export const StyledEstmSummCard = styled(StyledCard)`
-  &,
-  &:focus {
-    min-width: 21vw; /* slightly smaller cards in estimation summary */
-  }
+  min-width: 21vw; /* slightly smaller cards in estimation summary */
 
   @media ${device.desktop} {
-    &,
-    &:focus {
-      min-width: 80px; /* slightly smaller cards in estimation summary */
-    }
+    min-width: 80px; /* slightly smaller cards in estimation summary */
   }
 
   &,
@@ -165,11 +161,16 @@ export const StyledEstimationSummary = styled.div`
     margin-top: 0;
   }
 `;
+
 export const StyledEstimationSummaryList = styled.div`
   > span {
-    display: block;
+    display: flex;
+    align-items: center;
     margin-top: 8px;
     padding: 4px 0;
-    width: 50%;
+
+    @media ${device.desktop} {
+      width: 50%;
+    }
   }
 `;