+++ server/src/eventHandlers/includedInEstimations.js
@@ -1,10 +1,11 @@
 import {modifyUser} from './roomModifiers';
 
 /**
- * Sets "excluded" flag on user to false
+ * Sets "excluded" flag on user to false.
+ * (user is no longer "spectator")
  */
-const includedInEstimationsEventHandler = (room, eventPayload, userId) => {
-  return modifyUser(room, userId, (user) => ({
+const includedInEstimationsEventHandler = (room, eventPayload) => {
+  return modifyUser(room, eventPayload.userId, (user) => ({
     ...user,
     excluded: false
   }));