+++ CHANGELOG.md
@@ -1,3 +1,29 @@
+### 0.12.0: 2021-09-17
+
+* Display "Recommendation" in Estimation Summary: the first card that is larger than the numerical average. Eexcept numerical average is within 10% of the lower card. (#207, thanks to #ecaron)
+* You can now settle for any card. (Previously, only cards that were chosen by at least one user were possible settle values).
+* Dependencies updated
+
+### 0.11.6: 2021-09-04
+
+* Moved "New Round" button to the right. This prevents accidentally starting a new round, when clicking "Reveal" (#206, thanks to @ecaron)
+* Added autocomplete attribute to name input fields. Should prevent browsers from suggesting credit card numbers (#205, thanks to @ecaron)
+* Dependencies updated
+
+### 0.11.5: 2021-08-28
+
+* Modify "quick menu" on user's avatar: add option to quickly toggle "spectator" mode (exclude from estimations).
+* All users can now mark all other users as spectator.
+
+### 0.11.4: 2021-08-09
+
+* Ignore negative card values in calculating the "numerical average" (#192)
+* Dependencies updated
+
+### 0.11.3: 2021-05-08
+
+* Dependencies updated
+
 ### 0.11.2: 2021-03-26
 
 * Fix link to manual in auto-generated story (thanks to @kratz00)